# 8317720: RISC-V: Implement Adler32 intrinsic