… src/hotspot/cpu/riscv/vm_version_riscv.cpp