@ArsenyBochkarev 2024-04-09
@RealFYang Hi, thanks for pointing out! To achieve an additional acceleration, I did a vectorization and re-measured performance on Kendryte K230 with RVV 1.0 enabled:

Disabled intrinsic:

| Benchmark                     |         (count) |  Mode | Cnt   | Score |  Error   | Units |
| -------------------------------------- | ---------- | -------- | ------- | ------ | ------- | --------- |
| Adler32.TestAdler32.testAdler32Update |      64 | thrpt |  25 | 1867.257 | 10.034 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |    128 | thrpt |  25 | 1651.408 | 10.354 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |    256 | thrpt |  25 | 1345.505 |  4.847 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |    512  | thrpt |  25 |  976.550 |  3.889 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |   1024 | thrpt |  25 |  634.572 |  1.256 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |   2048 | thrpt |  25 |  371.763 |  0.588 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |   5012 | thrpt |  25  | 168.774 |  0.147 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |   8192 | thrpt |  25  | 106.578 |  0.135 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |  16384 | thrpt |  25  |  54.216 |  0.097 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |  32768 |  thrpt |  25 |   25.744 |  0.025 | ops/ms |
| Adler32.TestAdler32.testAdler32Update |   65536 | thrpt |  25  |  12.992 |  0.064 | ops/ms |

Enabled intrinsic:

| Benchmark                     |         (count) |  Mode | Cnt   | Score |  Error   | Units |
| -------------------------------------- | ---------- | -------- | ------- | ------ | ------- | --------- |
| Adler32.TestAdler32.testAdler32Update  |     64 | thrpt |  25 | 7177.572 | 13.724 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |    128 | thrpt |  25 | 4724.756 |  6.231 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |    256 | thrpt  | 25 | 2813.707 |  2.464 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |    512 | thrpt  | 25 | 1557.127 |  1.325 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |   1024 | thrpt  | 25 |  821.303 |  1.480 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |   2048 | thrpt  | 25  | 422.749 |  0.333 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |   5012 | thrpt  | 25  | 175.323 |  0.154 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |   8192 | thrpt  | 25  | 117.811 | 0.157 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |  16384 | thrpt |  25 |   58.990 |  0.081 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |  32768 | thrpt |  25 |   28.827 |  0.066 | ops/ms |
| Adler32.TestAdler32.testAdler32Update  |  65536 | thrpt |  25 |   14.773 |  0.116 | ops/ms |

It seems to me that there's a huge room for improvement in the current implementation.

BTW, the data I used as a comparison from T-Head board was recorded a few months ago. Is it the code generation that has improved significantly? Or it's just me making some kind of mistake in measurements?