@RealFYang 2024-06-14
Here you go. JMH data for the latest version:
```
Before:

Benchmark                      (count)   Mode  Cnt     Score    Error   Units
TestAdler32.testAdler32Update       64  thrpt   25  1839.455 ± 30.002  ops/ms
TestAdler32.testAdler32Update      128  thrpt   25  1633.132 ± 24.443  ops/ms
TestAdler32.testAdler32Update      256  thrpt   25  1312.842 ±  9.129  ops/ms
TestAdler32.testAdler32Update      512  thrpt   25   979.352 ±  8.982  ops/ms
TestAdler32.testAdler32Update     1024  thrpt   25   642.716 ±  0.414  ops/ms
TestAdler32.testAdler32Update     2048  thrpt   25   378.849 ±  1.095  ops/ms
TestAdler32.testAdler32Update     5012  thrpt   25   171.918 ±  0.479  ops/ms
TestAdler32.testAdler32Update     8192  thrpt   25   108.965 ±  0.084  ops/ms
TestAdler32.testAdler32Update    16384  thrpt   25    55.839 ±  0.185  ops/ms
TestAdler32.testAdler32Update    32768  thrpt   25    26.950 ±  0.069  ops/ms
TestAdler32.testAdler32Update    65536  thrpt   25    13.462 ±  0.020  ops/ms

After:

Benchmark                      (count)   Mode  Cnt     Score    Error   Units
TestAdler32.testAdler32Update       64  thrpt   25  7952.749 ± 18.472  ops/ms
TestAdler32.testAdler32Update      128  thrpt   25  6140.881 ± 74.067  ops/ms
TestAdler32.testAdler32Update      256  thrpt   25  4397.484 ± 32.142  ops/ms
TestAdler32.testAdler32Update      512  thrpt   25  2808.776 ±  0.749  ops/ms
TestAdler32.testAdler32Update     1024  thrpt   25  1636.976 ±  4.634  ops/ms
TestAdler32.testAdler32Update     2048  thrpt   25   891.656 ±  1.636  ops/ms
TestAdler32.testAdler32Update     5012  thrpt   25   382.059 ±  0.644  ops/ms
TestAdler32.testAdler32Update     8192  thrpt   25   237.452 ±  0.085  ops/ms
TestAdler32.testAdler32Update    16384  thrpt   25   120.433 ±  0.288  ops/ms
TestAdler32.testAdler32Update    32768  thrpt   25    58.892 ±  0.052  ops/ms
TestAdler32.testAdler32Update    65536  thrpt   25    29.891 ±  0.011  ops/ms
```